Removed modules imgeom and geocoding
